feat(base structure) - Creation of mixin, components, global variables and change of font-family.
Implementation of displayFlex mixin to help for fast use of flex-box
properties.
Creation of the global variables refering to the main colors pallette.
Creation of the ClassifiedAdsList and JobCard components along with it's
basic structure.
Change of default font-family from Roboto to Poppins.
Change of default font color from $GRAY to $BLACK.